Project Files/Desktop Apps/requirements.txt
Pillow==8.2.0
reportlab==3.5.67
ttkthemes==3.2.2
